requirements-render.txt
# Core Web Stack
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
uvloop==0.21.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-multipart==0.0.21
python-dotenv==1.2.1
sentry-sdk[fastapi]==2.22.0
fastapi-limiter==0.1.6
email-validator

# Database & Storage
sqlmodel==0.0.31
sqlalchemy==2.0.46
alembic==1.18.3
psycopg2-binary==2.9.11
cloudinary==1.42.2

# Auth & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0

# Clients (Offloading)
modal==1.3.2
huggingface-hub==0.36.0
groq==1.1.0
requests==2.32.5
httpx==0.28.1
aiohttp==3.13.3
langchain-core==1.2.7
langchain-ollama==1.0.1
websockets==16.0


# Utilities
jinja2==3.1.6
aiofiles==25.1.0
python-socketio==5.16.0
python-engineio==4.13.0
celery==5.4.0
loguru
tenacity==9.1.2
edge-tts==7.2.7
redis==5.2.0
aiortc==1.9.0
static_ffmpeg==3.0
orjson==3.11.5
pyyaml

# Image Handling (Headless & Minimal)
opencv-python-headless==4.13.0.90
numpy==2.3.5
pillow==11.3.0
pydub==0.25.1

# System Utils
psutil==7.2.1
cffi==2.0.0
cryptography==46.0.3
PyMuPDF==1.26.7
python-docx==1.2.0
openpyxl==3.1.5
pandas==2.3.3
